Bug: Submitted by:   "Daniel A. Steffen" <steffen@ics.mq.edu.au> Reviewed by: Approved by: Obtained from:
adding the "port test" target. This runs test suites bundled with a port.
For now, in order to have it run 'test.run' needs to be set to 'yes' in
the portfile (this was added to Daniel's patch, fenner@ suggested this).

while we're at it, portfile.7 has been fixed to no have any empty lines.

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@3268 d073be05-634f-4543-b044-5fe20cf6d1d6