Add dependency registration and query procs to the registry (registry.tcl and receipt_flat.tcl).  Add dependency registration to portinstall.tcl, and dependency checking to portuninstall.tcl.  This should provide fairly primitive dependency checking before allowing the uninstallation of a port that another port depends on (force in honored and will force the uninstall of a port even if other ports depend on it).
There are some problems with dependencies and Port Images still, in that deactivation breaks dependencies, since we depend on things being in ${prefix} rather than their image location.  I think this is about as good as it gets for now though, since Kevin plans to work on dependencies anyway (woot!).

git-svn-id: https://svn.macports.org/repository/macports/branches/wbb4-versionfoobage-1/base@5376 d073be05-634f-4543-b044-5fe20cf6d1d6