Changed the way the install phase is handled.
To do so, we create a new property for targets, target_state, which if no (defaults
to empty) means that the state is not added to the state file.

Install target now has this property and is no longer "always" for the runtype.

The consequence is that when the exact port/version/revision/variants is already
installed, all non-always targets are not performed and when such target are
skipped, the state file is not touched.

Additionally, if the user specifies -f or if the Portfile was modified since
installation, this behavior will be reverted. This is what we want intuitively.

Consequently, the following use cases now work:
- some port was deactivated and the dependency code wants to activate it, so it
will only activate them (all required phase will be skipped since the port was
already installed). This needed to be fixed and is the reason for this change.
- port install with a new revision on a deactivated port will work as expected.
port install with a newer Portfile will also work (useful for development). And
if the Portfile is older but the user really wants to install the port (-d says
why the steps were skipped), she can use -f. This is basically why the state
file is not modified. This is a side-effect of this change.

Please note that although all this ought to be documented, it should (partly) be
obsoleted by the new target sequence order, since changing the target order will
achieve the original goal of this modification (make dependency code
successfully activate ports).

Bug:
Submitted by:
Reviewed by:
Approved by:
Obtained from:

git-svn-id: https://svn.macports.org/repository/macports/branches/wbb4-versionfoobage-1/base@5569 d073be05-634f-4543-b044-5fe20cf6d1d6