Rollback change on set_ui_prefix. Apparently, there are still large parts of this source code that I don't fully grasp.
What happens is that ui_* from darwinports.tcl conflict with ui_* from
portutil.tcl and we cannot include both in this file. So we cannot call
set_ui_prefix since it's in portutil.tcl and we need darwinports.tcl for
this file.

We could add set_ui_prefix in darwinports.tcl, but I'm confused with that
name conflict.

Bug:
Submitted by:
Reviewed by:
Approved by:
Obtained from:

git-svn-id: https://svn.macports.org/repository/macports/branches/wbb4-versionfoobage-1/base@5574 d073be05-634f-4543-b044-5fe20cf6d1d6