Bug:            https://trac.macports.org/ticket/2162 Submitted by:   Eric Hall <opendarwin.org@darkart.com> Reviewed by:    wbb4@ Approved by:    wbb4@ Obtained from:
Switch to using registry::installed for dependency checking rather than registry::entry_exists.  The later is only designed to check for a specific version/revision/variant-combination of an installed port, and because dependencies currently do not understand versions, much less revisions and variants, this causes breakage (especially with revisions).  Using registry::installed allows us to simply check for a port name existing without modifying the registry code (which will be useful once dependencies are made better).

Thanks to Eric for being willing to wade into base/ and do this! ;)

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@8020 d073be05-634f-4543-b044-5fe20cf6d1d6