Some more enhancements to "archive mode":
1. Adds "-b" (binary-only) and "-s" (source-only) command-line options
to force a build from source, or to force only the use of binary
archives for installation. They only apply when portarchivemode=yes (or
is unset) in the ports.conf/.portsrc file. Feature suggested by Juan.

2. Adds some informational control files to the archives like
FreeBSD/Debian have (+COMMENT, +DESC, +CONTENTS). The DarwinPorts system
does not use these for anything at the moment, and they are subject to
change, but we may utilize them in the future for binary only
installation of ports without a "dports" tree present.

3. Adds "portarchivepath" option to the ports.conf/.portsrc file to
control where archives are written/searched for, it defaults to
${portdbpath}/packages. Feature suggested by Shantonu.

git-svn-id: https://svn.macports.org/repository/macports/branches/rshaw-archivemode/base@8376 d073be05-634f-4543-b044-5fe20cf6d1d6