Enhancement added to archive mode:
Multiple portarchivetype entries are now allowed and must be a colon or
comma separated list of choices (NO spaces).

Use of multiple types will cause archive creation to build all the
specified types in one step. This is useful for creating multiple
archive types during the same build step for comparison or distribution
to multiple machines that may not have the same archive tool support
(i.e. may not have xar, but have cpio).

Unarchive uses multiple types as a search list to locate the archive.
The first archive to match one of the specified types in the order
it is found is the one used.

git-svn-id: https://svn.macports.org/repository/macports/branches/rshaw-archivemode/base@8668 d073be05-634f-4543-b044-5fe20cf6d1d6