New feature added:
Adds "portautoclean" option to ports.conf, and equivalent "-c" and "-k"
command-line options. This new feature will automatically run "clean"
after any "install" of a port, and any "install" of a port's
dependencies. Note: Executing "port activate" will NOT run the clean
automatically. You must use "port install" for it to occur. This is so
that clean is not run when your just deactivating/activating a port.

The "-c" (clean mode) can be used to force autoclean mode ON, for the
case where you have portautoclean=no in your ports.conf.

The "-k" (keep mode) can be used to force autoclean mode OFF, for the
case where you have portautoclean=yes in your ports.conf (the default).

git-svn-id: https://svn.macports.org/repository/macports/branches/rshaw-archivemode/base@8832 d073be05-634f-4543-b044-5fe20cf6d1d6