Fixed StartupItem script creation to eliminate stray { ... } brackets when start/stop/restart command contained args like this: 	startupitem.start ${prefix}/bin/foobard -a -b This would cause { ${prefix}/bin/foobard -a -b } in the StartupItem script output.
Fixed start/stop/restart command parsing to properly substitute variables
to allow specifications like this (fix above also helps with this):
	startupitem.restart {
		${prefix}/bin/some command line
		${prefix}/bin/another command line
	}
And still properly evaluate the ${prefix} variables.

Fixed to properly handle startupitem.requires items also, so they can
be easily specified like this:
	startupitem.requires	Portmap DNS Whatever
and it will properly output the line in the plist like this:
	Requires = ("Portmap", "DNS", "Whatever");
before, it would have had { ... } around the values when more than one
word was in the value and/or you'd be required to put backslashes and
commas in the value which is ugly, and not very readble in a Portfile.

RCng script setup was NOT touched, as I don't believe the above items
would work and/or are not necessary. If needed, similar fixes could be
made, but I'll leave that up to Landon (or someone else more versed in
RCng to do that, if it's needed).

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@10678 d073be05-634f-4543-b044-5fe20cf6d1d6