Removed subst calls for fix that was supposed to allow syntax like this:   	startupitem.restart {   		${prefix}/bin/some command line   		${prefix}/bin/another command line   	} So that it would properly evaluate the ${prefix} variables. Well, it does for prefix, and any variable in the startupitem function itself, but no others, so it didn't work quite right if you used something like ${myvar} that was defined in the Portfile. DOH!
So to make the above really work, we have to suffice with:
  	startupitem.restart "
  		${prefix}/bin/some command line
  		${prefix}/bin/another command line
  	"
So that it evaluates in the Portfile context. If anyone else knows how
to properly evaluate the above case correctly in the Portfile context,
let me know. I like the syntax of { ... } instead of the " ... " since
you don't have to escape "'s in the commands then.

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@10680 d073be05-634f-4543-b044-5fe20cf6d1d6