Submitted by:   pguyot@ Reviewed by:    jmpp@
Implementation of a RAM only based filemap.

NOTE: This commit compliments the filemap concurrency bugfix of the previous commit
with pending patches to the filemap.c file and tests/filemap.tcl, as the bug was
actually present in the ram based filemap. Needless to say, this commit should have
come before!

git-svn-id: https://svn.macports.org/repository/macports/branches/release1/base@13568 d073be05-634f-4543-b044-5fe20cf6d1d6