Submitted by:   pguyot@ Reviewed by:    jmpp@
Small fix for the Xcode GroupCode to make it default to -all targets, plus
the addition of the group's full documentaion.

NOTE: I would personally like to see the xcode.7 page being moved to less
xcode specific and more GroupCode wide open portgroup.7 page, in which we
can eventually include the documentation of all the groups we currently support.
Also, I believe the xcode.7 name is misleading as it can confuse people into
believing the page belongs to the Developer Tools package. I am backed in this
proposal by Ole and others (Ernie and Charlie Allom, for instance), so in any
case I'd like to hear a valid argument *against* the rename.

git-svn-id: https://svn.macports.org/repository/macports/branches/release1/base@13970 d073be05-634f-4543-b044-5fe20cf6d1d6