Submitted by:   jmpp@
I'm finally starting the recovery of the "Available Ports" site, let it be
permanent this time! I plan to have a dedicated "dpadmin" (or anything to that
effect) OD based account that will take care of the job of keeping the page's db
up to date on an automated fashion, along with other chores such as the index
regen one, which is now being taken care of by Paul's account on Lamancha
(which shouldn't be). The appropriate request will be routed through admin@

In the mean time, I suppose there's no reason why management scripts in
base/src/portmgr/ shouldn't be installed, so here's a minor abstraction to
the PortIndex2MySQL.tcl one to make it work with non-default dp installations.
Other scripts can be adapted later for proper installation too.

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@14700 d073be05-634f-4543-b044-5fe20cf6d1d6