Change to port(1) syntax:
	- @ is now required before any version number, in order to disambiguate
	  cases where the portname starts with a digit, or the version doesn't.

	- The fields of a portspec are now uniquely identified by their
	  leading character:

	  	- The version must begin with an @
	  	- A variant must begin with a - or +
	  	- An option must have a c-style identifier followed by an = sign
	  	- A portname is none of the above

	- The fields may be combined into a single argument, if some rules are
	  followed:

		- Options must always be in an argument by themselves
		- The version, if present, must always follow the portname
		- If no version is present, then any variants must be in
		  subsequent arguments.

That sounds more complex than it really is. Some examples will illustrate:

Given a port named "foo" and variants "v1" and "v2" and a version 1.0,

The following are legal:

	port echo foo @1.0 CFLAGS=FOO LDFLAGS=BAR
	port echo foo +v1
	port echo foo +v1 - v2
	port echo foo@1.0+v1-v2
	port echo foo @1.0+v1-v2
	port echo foo +v1-v2

The following are not legal:

	port echo foo+v1		(no version is present, so variants must follow)
	port echo foo 1.0		(@ must introduce the version)

Note that the portname may be omitted, and will default to the port in the current
directory, but only if it is the first port named (else it wouldn't be distinguishable
from the preceeding port). The pseudo port current may help you out in this case:

	port echo @1.0				(current port, at version 1.0)
	port echo a @1.0			(port named a at version 1.0)
	port echo a current @1.0	(port named a and current port at version 1.0)
	port echo a@ @1.0			(this should also mean port a and current port at version 1.0)


I changed some of the version reporting (installed, list, echo) to use the @ syntax
for version so that it gets people on the right page. We may need to do this in a
few more places as well. Let me know if you see some.

Feedback welcome.

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@14847 d073be05-634f-4543-b044-5fe20cf6d1d6