Submitted by:   jmpp@ joe@ jberry@
Finally, the main documentation reference for our 1.2 release. Incomplete as
it still is, it is nevertheless "complete enough" to go and, thus, make up
for one the final patches that will go into the 1.2 sources before we declare
a final freeze and release... real soon now!

Further enhancements to this page (many of them already in store) will come
at a later date maybe as point releases against 1.2

James, Joe, thanks a lot for your help and constant reviews to my edits, indispensable
to complete such large enhancement against what we had in 1.1. Hope people like it!

git-svn-id: https://svn.macports.org/repository/macports/branches/release_1_2/base@15484 d073be05-634f-4543-b044-5fe20cf6d1d6