Submitted by: joe@ jmpp@
Final ReadMe.rtf file for our pkg, that we now build through DP itsel, woot!

This should now make up for the *final* commit to the 1.2 sources before we release,
unless there's something huge we forgot that can't wait for a point release (anything?).
So, in short, we should be good to go.

Thanks for this great looking file Joe!

PS: I have to wait for the final 1.2 source tarball to be posted to commit this to the
darwinports files/ dir (along with the needed changes to the corresponding Portfile).

git-svn-id: https://svn.macports.org/repository/macports/branches/release_1_2/base@15486 d073be05-634f-4543-b044-5fe20cf6d1d6