Submitted by:   jmpp@
Per Paul's advice, downgraded a possible chmod failure to a warning,
rather than erroring out on it.

On the other hand, overlooking jberry's recent advice to pass -f to chmod(1),
as I believe we should catch and warn the user if parts of his/her dports tree
are not world readable.

All and all, this is still just a hack, if anyone has a better fix (the issue
being that in some situations the dports tree might not be readable to the user,
like after sync'ing for the first time with a quite restrictive umask) please
do feel free to step in and contribute. Keeping in only in HEAD for the moment,
maybe up for a future 1.3.2.

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@18822 d073be05-634f-4543-b044-5fe20cf6d1d6