What was supposed to be a small commit to the InexRegen.sh script turned into a somewhat major rewrite! Changelog, in order of relevance:
 * First and foremost: We now commit the index even if any number of ports fail parsing (I figure that at least
   an incomplete index is better than no index what-so-ever for a full 12 hour period);
 * Indexing failures, if any, are grep'd for from the runtime log and appended to the commit messsage;
 * Renaming the TREE variable to SRCTREE for clarity;
 * The ${SRCTREE}/dports and ${SRCTREE}/base trees are now checked out if non-existent, updated otherwise (we now
   don't cowardly bail out if they don't exist)
 * Introduced a bail() function that handles mailing the runtime log upon failure and, namely, bailing out;
 * Removed the now unnecessary FAILED variable and all the chekcs for its value;
 * Introduced a cleanup() function that takes care of deleting the runtime log, commit message file and lock;
 * Innocuous DP --> MP move;
 * Somewhat improved error reporting;
 * And, last but certainly not least, syntax re-tabbing! (Sorry guys, couldn't help it!!)

I tested this new script with an environment as close to reality as possible and it seems to be working dandy,
but some fresh eyes on it other than mine and real life testing are surely necessary before putting it to work.
Daniel, here's where you come in! ;-)


-jmpp

PS: Lets work on this file on the release_1_4 branch, as that version is the one pulling the base sources we
are currently indexing with in real life. Once it's done I'll merge all of its commits (if more than one) back
into trunk.

git-svn-id: https://svn.macports.org/repository/macports/branches/release_1_4/base@23256 d073be05-634f-4543-b044-5fe20cf6d1d6