Prematurely commiting a new test proc, "findBinary", that will eventually unite how our base code finds external programs we need. It's meant to eliminate direct calls to both the "binaryInPath" proc and autoconf module variables to determine the path to binaries, concentrating them all in unified calls to this new proc.
This is still pretty much experimental code and does not yet work as expected. The "file" and "rsync" synchronization methods in the
mportsync proc have been adapted to use "findBinary" and no longer do their business, so clearly more work is needed. However, it is
my intention to extend "findBinary" usage to the rest of our sources once bugs have been ironed out.

git-svn-id: https://svn.macports.org/repository/macports/branches/dp2mp-move/base@24949 d073be05-634f-4543-b044-5fe20cf6d1d6