Port sources are now downloaded into a ${prefix}/var/macports/sources/<server>/<tree_offered> skeleton. In the most common scenario, our default configuration, server = rsync.macports.org and the offered tree is release/ports (new layout from this branch), so the full path will most likely translate into /opt/local/var/macports/sources/rsync.macports.org/release/ports.
I believe this approach is both intuitive and adaptative: on the one hand, it allows us to
organize the fetched sources (either base or the ports tree) according to the server that
offers them (thinking about mirrors in the future) and, therein, according to quality
qualifiers used by the vendor, e.g. release, stable, unstable, etc.

git-svn-id: https://svn.macports.org/repository/macports/branches/dp2mp-move/base@25158 d073be05-634f-4543-b044-5fe20cf6d1d6