This script is only meant for users with a fully default environment, and in such bash is the default shell so let's not glob for all bourne-based shells. We deal with tcsh too by extension because that was default for Mac OS X at one point, but in a similar fashion let's limit ourselves only to that shell and not all C-based shells with the *csh glob.
This commit follows advice from Vincent Lefevre on the dev list that modifying the environment
in ~/.profile for the zsh shell as we would here, thanks to the *sh glob, is ill-advised. What I
think about that is: a) anyone running zsh is *not* running a default environment and thus is
outside the scope of this script; b) anyone savvy enough to switch to zsh (or to any other shell
away from the default, for that matter) should also be savvy enough to set up his/her own environment.

git-svn-id: https://svn.macports.org/repository/macports/branches/release_1_6/base@31818 d073be05-634f-4543-b044-5fe20cf6d1d6