Even better than https://trac.macports.org/changeset/32968: pass the new setting that we want to write to the custom 'write_setting' command as ${2}, which allows us to get rid of that case construct within it and thus truly abstract it away.
Now 'write_setting' takes two arguments: the environment variable that is to be tweaked (${1}) and how it should be
tweaked (${2}), which makes a lot more sense. 'write_setting' itself doesn't need to be adapted any more in case the
settings need to be changed, new ones added or even some deleted.

git-svn-id: https://svn.macports.org/repository/macports/branches/release_1_6/base@32970 d073be05-634f-4543-b044-5fe20cf6d1d6