* Use dscl(1) to detect the default shell for the installing user, rather than trying to infer it from the inherited environment    (which did not reflect the user's at all, as it is entirely imposed by Installer.app when the pkg is used in real life). Thanks    to Ryan on macports-dev for the original suggestion!
 * Provide a little more information about the shell we detected.

NOTE: I tested this edited script through a test pkg on Leopard and it works, both when used from an account with a bash shell and
from one with a tcsh shell. If anyone cares to test on Tiger and/or Panther, they should contact me so I can hand them the pkg, thanks!
In any case, I'm feeling confident it'll work consistently this time round to detect the shell type, given the nature of the dscl operation.

git-svn-id: https://svn.macports.org/repository/macports/branches/release_1_6/base@33639 d073be05-634f-4543-b044-5fe20cf6d1d6