Added port stats action.
This action takes one optional argument - submit

If no argument is given then port stats displays (in human readable form) the 
information that will be submitted to MacPorts. As of now only information 
relevant to building ports is displayed.

The submit subcommand will initiate the actual upload. Right now it exists 
only as a placeholder and outputs the text "Will submit collected data"

git-svn-id: https://svn.macports.org/repository/macports/branches/gsoc11-statistics/base@78936 d073be05-634f-4543-b044-5fe20cf6d1d6