Use xcrun to discover all compilers supplied by xcode.
Xcode 4.3 previews continue to not install all compilers in /usr/bin,
so we call xcrun to ask it where these compilers are. This should
continue to work if or when xcode builds actually do install in
expected locations, and shields us from this sort of change.

xcrun --find works back into the xcode 3 days as far as I can tell.

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@88777 d073be05-634f-4543-b044-5fe20cf6d1d6