Adjust compiler defaults on old OS versions
Removes apple-gcc-4.0 from the list as it barely works.

Moves the XCode 3.0/3.1 (Leopard) default to gcc-4.2 to make it more
similar to better supported OS versions.

Adds apple-gcc-4.2 as a first-choice fallback on Tiger.

Adds macports-clang-3.0 to the fallbacks on Tiger and Leopard
(although they are only well tested on Intel, so they are listed
after apple-gcc-4.2)

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@92109 d073be05-634f-4543-b044-5fe20cf6d1d6