Use macports-llvm-gcc-4.2 as a backup instead of llvm-gcc-4.2 with XCode 4.6
It was announced that llvm-gcc is deprecated in XCode 4.6 and will be removed
in the next major release.  This updates base to fallback on the llvm-gcc42
port rather than the Apple-provided llvm-gcc when using this version of XCode
to allow developers using base trunk to find issues with ports that may exist
with the removal of llvm-gcc from XCode.

I do not reccomend shipping this change in the 2.1.x release series.

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@102258 d073be05-634f-4543-b044-5fe20cf6d1d6