portpkg::pkg_main: in log message: separate port name from version number.
This will make it easier to parse log files for the port name and the
log file.  For example, the previous message style:

  Creating pkg for fftw-3-0_3.3.3_1

and the new style:

  Creating pkg for fftw-3 version 0_3.3.3_1

The later is easier to see that the "-3" is part of the port name and
it makes it easier to use awk to get the desired fields.

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@110692 d073be05-634f-4543-b044-5fe20cf6d1d6