Use autoconf.args instead of autoconf.pre_args
The arguments passed to the autoconf/autoreconf/automake commands are likely
depending on the .cmd used. Therefore, it's unlikely that we need additional
commands, but more likely we need completely different or no arguments (for
instance when running ./autogen.sh).

This change might break some existing ports that clear autoconf.pre_args which
can be fixed by clearing autoconf.args as well. Applying this to a Portfile is
backward-compatible.

https://lists.macosforge.org/pipermail/macports-dev/2013-September/024401.html

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@113062 d073be05-634f-4543-b044-5fe20cf6d1d6