portconfigure.tcl: Add -stdlib to configure.{cxxflags,objcxxflags}
The configure.*flags options provided to Portfiles don't reflect the
final values assigned to the corresponding *FLAGS environment variables.
The configure phase adds several additional items to the environment
variables, including "-pipe", "-stdlib", "-isysroot", and "-arch" flags.

As discussed on macports-dev, it would be useful for configure.*flags to
contain the exact string that will go into the environment (to use with
build.args, for example).

As a first step, ensure that the "-stdlib" flag is present in
configure.cxxflags and configure.objcxxflags, as necessary. As
a corollary, any Portfile-supplied "-stdlib" flags are deleted with
extreme prejudice, so, uh, just don't put "-stdlib" in
configure.cxxflags.

https://lists.macosforge.org/pipermail/macports-dev/2014-October/028279.html

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@126356 d073be05-634f-4543-b044-5fe20cf6d1d6