base: darwintrace: close sockets when threads are terminated
If a long-running process started and terminated a large number of processes,
the file descriptors the were opened for these threads would not get closed.
Instead, they would linger around in memory, unreferenced, until the process
was terminated completely.

This can cause exhaustion of file descriptors and fail builds. I've seen this
behavior with GHC.

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@138530 d073be05-634f-4543-b044-5fe20cf6d1d6