base: pextlib/system: Use exit(3) to flush buffers
_exit(2) doesn't flush buffers that may be still in use, such as output printed
to stdout. Use exit(3) to simplify reporting a problem in a subprocess
pre-execve(2).

Additionally change the return value in case of a failed exec to 128 to signal
a failure in a subprocess; this is a convention often used by GNU software, and
I think it makes sense to help distinguishing this from a failure in the
executed program.

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@141414 d073be05-634f-4543-b044-5fe20cf6d1d6