base: pextlib/system: Report signal termination
Add code to correctly report termination by signal back to the Tcl user. See
https://www.tcl.tk/man/tcl7.6/TclCmd/tclvars.n.html#M7 for a description of the
format in case of signaled subprocesses. See vendor/tcl8.5.15/generic/tclPipe.c
for C code that generates these messages. Use Tcl_SignalId and Tcl_SignalMsg to
generate the required data for the message to be returned to the user.

git-svn-id: https://svn.macports.org/repository/macports/trunk/base@141415 d073be05-634f-4543-b044-5fe20cf6d1d6