Support multi-valued maintainers
Since our move to GitHub it is sometimes hard to find out whether a pull
request was sent by a maintainer or what a maintainer's Trac or GitHub
account is. This can be solved by allowing GitHub usernames as
maintainers, but puts us in the situation of not having an email address
on file for a maintainer.

Solve this by supporting multi-valued maintainer fields using Tcl lists,
so that

 maintainers {@github-username macports-handle example.com:localpart}

works and is displayed as beloging to a single person.

While we're at it, drop the two implementations of unobscure_maintainers
and provide a public API in macports1.0 to be used.

Additionally, add tests that verify the behavior of
macports::unobscure_maintainers.