doc: Obtain manpage timestamp from Git
The date information written into our manpages used to be obtained from
the SVN $Id$ tag placed in the files. Now that we have switched to git
and those tags end up empty, the file date is automatically used
instead.

See http://www.methods.co.nz/asciidoc/userguide.html#X95 where this is
documented.

Since Git does not preserve file modification dates on checkout or pull,
using the file modification time will yield different results on each
developer's machine. On the other hand, we should support building the
manpages when the source code is extracted from a tarball (where using
the file modification time is probably fine).

Solve this by explicitly passing a time with `-a revdate`. The timestamp
is obtained from Git, if available. The fallback is to set the `revdate`
attribute to empty, which will cause the file modification time to be
used.