Update git tarball instead of writing a new one
The archive tarball created by git has a pax header comment referencing
the sha1 of the commit it was created for (see `git get-tar-commit-id`).
This would be lost when moving everything to a new tarball. In the rare
case that the submodule command did not actually create any new tarball
content, this will now leave the original git tarball unmodified.