Use a 'catch' wrapper that can ignore signals
The addition of Tclx signal handlers means that SIGINT and SIGTERM
now result in an error being raised. In most places where the catch
command is used, we don't actually check if the error is a signal, and
in most of those cases can't do anything sensible except re-raise the
error anyway.

Thus we replace 'catch' with our own version which encapsulates
checking if the error is a signal and (by default) re-raises it if so.
The -signal flag can be passed as the first argument to the new catch
to make it catch signals too.

The code is placed in a new package 'mpcommon' so it can be used via
'package require' in both the main interpreter and the slave interpreters
that execute Portfiles. It can't simply be aliased in to the slaves like we
normally do for shared commands because it uses uplevel and aliases add an
extra call level.