Remove altprefix support
This feature allowed using an alternative $portdbpath under ~/.macports
when port was running with insufficient privileges to use the normal
one. This added complexity and didn't really have a good reason for
existing, except as a way to test that ports didn't bypass the destroot
(as they would fail to install files directly into $prefix when running
without privileges). With the addition of sandboxing, this reason went
away, so now the code can too.