Fix the detection of /opt/local/bin in user's PATH (#16)
The postflight script was using the SHELL variable, which is root's shell
and not USHELL, which is the user's shell. Use full path of the user's
shell for USHELL.

The test for MacPorts paths in the users PATH can be made more robust.
The old code will match on /opt/locale for example.

Spit the PATH into one path per line with tr and then grep for an exact match
with /opt/local/bin.
(cherry picked from commit dcb07880b7e8a47a62ddd09b29ed00d02a373e82)