mportselect: always delete existing symlinks
If the previously selected port was uninstalled, the 'current' symlink
would be left pointing to the old select file that is no longer
available.

The condition using 'file exists' follows the symlink and checks its
target, not the symlink itself. It is safe to always call 'file delete'
as deletion of a non-existing target is not an error.

Closes: https://trac.macports.org/ticket/31702
(cherry picked from commit 9389e838454b8ad9792c41a5128ff1178f9b2542)