misc src/: switch from "string is integer" to "string is entier".
The "string is integer" check bails out of the passed data does not fit
in an 32bit (signed) integer data type. Contrary to that, "string is
entier" has bignum support and no constraints on the actual value.

All this said, sqlite3 might fail if the value doesn't fit into a signed
8-byte-integer, but we don't actually have a recovery plan for that
yet... and beside `port lint`, we don't check for such an overflow in
the first place.

Fixes: https://trac.macports.org/ticket/53489