Port diagnose PATH check improvements
Get shell from dscl instead of assuming bash. Don't try to check the
value of PATH with shells we don't know about. The code to add our dirs
to PATH only works with bash, so don't offer to do it with other
shells. Use the interactive questions API instead of reimplementing it.

Fixes: https://trac.macports.org/ticket/53473