misc: let 'port provides' and trace mode honor the FS case-sensitivity.
Extend Pextlib with functions to fetch the file system case-sensitivity
value for a specific path. The TCL interface than uses this function
when mapping a file path to a port name is requested.

cregistry and registry2.0 were updated to perform lookups of paths to
port IDs or names case-insensitively if requested by using the "COLLATE
NOCASE" option.

For this to actually work in a timely manner, a new "COLLATE NOCASE"
index is needed for the actual_path column. Thus bumping the internal DB
version number.

This fixes bugs where port build systems access files on
case-insensitive file systems with different capitalization than
registered in our database - thus bypassing trace mode's sanity checks
as the files in question are "unknown".