Remove checks for ObjC runtime and foundation
After removal of cflib and tclobjc, there is no code requiring
a Objective-C runtime or oundation libraries in base anymore.

The autoconf macros PTHREAD_LIBS or PTHREAD_CFLAGS are not used and have
no other references in the configure script. Therefore this is now also
unused code and can safely be removed.