Implement MacPorts migration
When an macOS install running MacPorts is upgraded, MacPorts will
require a rebuild of itself and reinstallation of all ports as
a robustness measure to ensure all ports continue working (and
compiling) as expected.

This commit is a squash of the work done by Umesh Singla during Google
Summer of Code 2017 for MacPorts, implementing a MacPorts command 'port
migrate' that automates the entire process of reinstalling MacPorts and
re-installing all installed ports. It was squashed to simplify rebasing
the history against the current master.

In addition to 'port migrate', this also adds two new commands 'port
snapshot' and 'port restore', which are the plumbing below migration.
Migration snapshots are stored in SQLite.