sip_copy_proc: Use copyfile(3) in sip-workaround
Use copyfile(3) to copy binaries to the sip-workaround location. If the
filesystem supports it (such as APFS), the flag COPYFILE_CLONE is
supposed to share the underlying filesystem blocks as copy-on-write,
therefore reducing required disk space and runtime. If this is not
possible, it will just fallback to a normal copy operation.

copyfile(3) is available on Mac OS X >= 10.5 Leopard. The previous
implementation will be used for compatibility with older releases.

(cherry picked from commit 3d4c9b342d28abd0b7aaf7eb70fa4862e898542c)