Fix sometimes leaving orphaned Portfiles in the registry
Portfiles are stored per subport,version,revision as there is no field
in the registry for the name shared between subports. Therefore, before
deleting a Portfile, we need to check only that the same subport is not
installed with different variants from an identical Portfile, not
whether *anything* else has an identical Portfile, since we can only
delete the one for this particular subport.