Adjust what modelines are warning and info.
Only warn if a modeline is found that is not on the first line.
The guide has conflicting info as to whether a modeline is required
and what it should be.

"This should be the first line of a Portfile. It sets the correct editing options for vim and emacs. See Port Style for more information. Its use is optional and up to the port maintainer."
"By default, the top line of all Portfiles should use a modeline that defines soft tabs for the vim and emacs editors as shown."