vendor: Fix permissions of tcllib on install
The tcllib installer uses a plain Tcl 'file copy' and not install(1) to
install its files and does not set explicit permissions. Therefore the
installed files will inherit the permissions of the files in our source
tree. To ensure the installed files are always readable by everyone as
expected, manually apply permissions with chmod after installing.