tracelib: Use binary search in dep_check()
Keeping the list of dependencies sorted allows using binary search when
determining whether a file is provided by a port in the dependency tree.

Depending on the length of the list of dependencies, this can yield
build time improvements of around 10 %.

Additionally, use Tcl's list functions to interpret the list of
dependencies rather than splitting at whitespaces manually.