migrate: Upgrade base first, automatically restart
In case a macOS upgrade breaks uninstallation of ports or using the
database, we should upgrade MacPorts base before saving the list of
ports and starting port migration.

Move rebuilding MacPorts to the front of 'port migrate', guard it with
a check indicating whether a rebuild is necessary and add user
interaction questions for the base rebuild. Deal with the case of users
with a pre-release version of MacPorts installed and avoid trashing
their installations.

After base was upgraded, automatically re-invoke 'port migrate' with the
'--continue' flag that will pick up migration where it left off. If
'port migrate' is invoked without '--continue' (e.g. when a user
manually recompiled their base installation), ask for confirmation
before uninstalling all ports.

Change selfupdate to also update the ports tree just in case users have
not used MacPorts in a while and the new ports tree contains build fixes
for the new macOS release.