Support startupitems that aren't generated by base
Whether a port has a startupitem is now indicated by startupitem.type
not being "none". Setting startupitem.create indicates that base should
generate a startupitem, and changes the default value of
startupitem.type.

If you want a traditional startupitem to be generated, nothing changes.
If you want to provide your own, set startupitem.type to the type you
are providing and install it yourself; and don't set
startupitem.create.

Fixes: https://trac.macports.org/ticket/36770