Store C++ stdlib info in the registry
There are two new registry properties for ports:

cxx_stdlib - this can be either "libc++" or "libstdc++" for most ports
that use C++, or "none" for ports that don't. It can also be "mixed"
for ports that end up linked to both runtimes for whatever reason.

cxx_stdlib_overridden - this is 1 if the port changed configure.cxx_stdlib
from its default value, 0 otherwise.

Since we need the info to figure out stdlib linkage, the 'binary'
registry attribute for files is now set at install time.

All this new info is also stored in the archive metadata.