doc: Drop modification time from manpages
Replace the manpage modification date with the MacPorts version. This
allows us to remove the extra call to git to find the last commit date,
and thus no longer requires us to commit the .txt file and the generated
manpage in separate commits.

Additionally, this removes the possibility of the date going backwards
when merging older pull requests with an old author date. We can also
not use the commit date to generate the date, because we usually rebase
commits when merging (which changes the commit date).