tests: Avoid bogus error message during test runs
The distcheck unit test caused remote network traffic, which is bad from
a testing perspective. Additionally, it printed error messages to
stderr.

Avoid both of these problems by replacing the curl command with a mock
for the time of the test. This will also speed up the test.

Drop the unit test for the filesize distcheck.type, since there is no
good way to actually test whether it succeeds.

Closes: https://trac.macports.org/ticket/52218