tests: Fix test failure due to missing SVN command
Remove the dependency on the SVN command for the svn-and-patchsites
regression test by setting svn.cmd to true.

Additionally, do not fetch vim patchfiles from a remote location, but
just use a file:// URL for the patchfiles.

Refactor the test to show the error output immediately if it fails,
rather than a string indicating that there was an error without the
actual output.