fetch: Switch tarball from xz to bzip2 compression
Although xz is more modern and would produce smaller files, it is not
available on macOS by default. Therefore we would have to first install
xz even if we would not necessarily use it for the fetch. Switch to
bzip2 instead which is always available and does not need the fetch
dependency.