fetch: Remove custom *.args for VCS fetch.types
Additional options could interfere with tarball creation and defeat the
goal of reproducibility.

Furthermore, the non-standard fetch routines for VCS need to run
multiple commands now. It would not be clear if the passed arguments
should only be used for the checkout/export command or also for others.
If there is a need for special cases, we should add specific options to
the fetch.types in question to control additional aspects.