mportexec: Remove misleading errorInfo output
The port target was evaluated in a separate Tcl slave interpreter, which
depending on the way it exits, does not necessarily update $::errorInfo
of the master interpreter. Therefore $::errorInfo in the master
interpreter often contains unrelated error information which leads to
confusion in the log file and caused tickets to be filed with the wrong
information.

All error information related to the failure should have already been
emitted by the slave interpreter.

Closes: https://trac.macports.org/ticket/55560