Allow multiple startupitems per port
The portfile interface is a new option called 'startupitems'. This
should be filled with a list of key/value pairs, with the names of the
keys corresponding to the startupitem.* options. Each startupitem
declared this way must speficy at least a name. Other options will be
taken to have the value of the corresponding startupitem.* option if
not present. Options apply to the name that appeared immediately prior.

Example:

startupitem.create  yes
startupitems        name        mydaemon \
                    executable  ${prefix}/bin/mydaemon \
                    name        anotherthing \
                    start       "${prefix}/bin/another --someflag"
                    stop        "${prefix}/bin/another shutdown"

Fixes: https://trac.macports.org/ticket/44295