mportopen: Normalize file:// porturls
The evaluation of some targets may switch the current working directory,
which means a relative porturl will no longer point to the Portfile we
used. Normalize relative paths in a file:// porturl in order to be
able to calculate the correct path for the _resources directory in the
corresponding ports tree.

Update a macports_dlist test case which relied on a "file://." porturl.

Closes: https://trac.macports.org/ticket/54743