Support startupitems installed as LaunchAgents
Distinguish beteeen LaunchDaemons and LaunchAgents based on
startupitem.location. Require root privileges to load/unload Daemons,
but load Agents as $sudo_user if available and running as root, or the
current user if not root.

Check if LaunchAgents are loaded by looking in the appropriate user's
domain, as opposed to the system domain for LaunchDaemons.