selfupdate: Use standard_configure.sh
This makes the MacPorts base you get after selfupdate more similar to
the MacPorts you get after using the binary installer.

standard_configure.sh enforces a PATH that does not contain the MacPorts
prefix, so there's no longer a need to set CC to guard against a
compiler in the MacPorts prefix.

standard_configure.sh defaults to enabling readline, so add a flag to
disable readline when needed.