Add script to set cxx_stdlib_overridden on upgrade
You can't tell whether configure.cxx_stdlib was overridden by a port if
it was not recorded at install time. This creates a problem for users
upgrading from versions of MacPorts that did not record it: rev-upgrade
will think that ports that do override it are broken, because they do
not follow the global cxx_stdlib setting.

Therefore, to ease the transition, keep a list of ports that are known
to override the setting, and when upgrading base, set the flag in the
registry for those ports if it is currently null.

Closes: https://trac.macports.org/ticket/56326