Change "Mac OS X" to "macOS"
This is not correct for older systems, where the historical name should
be used. The MacPorts port could be modified to replace this correctly,
but because I plan to overhaul the installer and move its creation from
the MacPorts port to a Makefile, I will postpone addressing this issue
until later.