Fix a crash due to an out-of-bounds array access
entry_search had a "type confusion" bug where it would perform array
accesses that bypassed bounds checks if given arguments that looked like
strategies. I've rewritten most of the argument parsing loop in a way
that should eliminate this issue, while being shorter and more
straightforward (with less duplicated code), only requiring one pass
through the arguments. The downside is that we allocate a bit more
memory and have a couple of gotos to handle errors.