Cache locations and versions of compilers
This improves worst-case performance considerably. This is mainly
because asking xcrun where a tool is takes a while when that tool is
not present on the system, and we did this multiple times. In one test,
parsing a single Portfile took 10 seconds, reduced to 0.5 by this
change. The benefits are even greater when multiple Portfiles are
parsed in a single invocation.