Allow falling back to "macosx" SDK
In the past, we could use the Command Line Tools as effectively an SDK
installed in /. That's no longer possible starting with 10.14, so when
an Xcode version is released that will run on 10.14 but doesn't come
with a 10.14 SDK, we'll break. This works around that, though it comes
with its own set of problems.

Closes: https://trac.macports.org/ticket/57143