macports1.0: Only check Xcode if building from src
We don't need to check the Xcode version (or presence) if we're not
going to build a port from source. This was previously achieved with
a shortcut in _check_xcode_version that would always return 0 if an
archive was available.

This fixes a situation where _check_xcode_version would indicate Xcode
isn't required because an archive is available even though a user
specifically asked for a build from source (e.g. when running 'sudo port
destroot' on a port that sets `use_xcode yes`).

The newly introduced _target_needs_toolchain now returns yes for all
cases where a toolchain is required regardless of archive availability
and returns yes iff an archive is available for those targets that only
need a toolchain when no prebuilt archive is being used.

See: https://github.com/macports/macports-base/pull/140