gcc_min_version: always allow macports-gcc on PowerPC
on Intel systems, macports-gcc is allowed only if the stdlib selection allows it.

on PowerPC, we have to be more flexible and allow macports-gcc versions
even if there might be a stdlib issue, as there are often no other options.

Luckily no issues have been raised with this in the several years it has been allowed.

This selection can be revisited if & when a suitable clang compiler and libc++
implementation become available on PowerPC